backend/pyproject.toml
[tool.poetry]
name = "uae-digital-identity-platform"
version = "0.1.0"
description = "UAE Digital Identity Platform Backend"
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
redis = "^5.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
structlog = "^24.1.0"
python-multipart = "^0.0.9"
email-validator = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.9.0"
ruff = "^0.3.4"
pre-commit = "^3.6.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "B", "I"]
